update styling on update user
