+++ controllers/api/users.js
@@ -2,16 +2,17 @@ const User = require("../../models/user");
 const jwt = require("jsonwebtoken");
 const bcrypt = require("bcrypt");
 const uuid = require("uuid");
-// import boto3 from "boto3";
-S3_BUCKET = 'gacatcollectordjango'
-S3_BASE_URL = 'https://s3.us-east-1.amazonaws.com/'
-const REGION = "ca-central-1"
+const S3 = require("react-aws-s3")
+const config = {
+  bucketName: process.env.REACT_APP_BUCKET_NAME,
+  region: process.env.REACT_APP_REGION,
+  accessKeyId: process.env.REACT_APP_ACCESS,
+  secretAccessKey: process.env.REACT_APP_SECRET,
+  S3_BASE_URL:process.env.S3_BASE_URL
+}
 
-// Source: https://javascript.plainenglish.io/how-to-upload-files-to-aws-s3-in-react-591e533d615e
-// const myBucket = new AWS.S3({
-//   params: { Bucket: S3_BUCKET},
-//   region: REGION,
-// })
+
+ 
 
 async function create(req, res) {
   try {
@@ -48,23 +49,16 @@ async function updateUser(req,res) {
     console.log ("error in update")
   }
 
-  photo_file = request.FILES.get('photo-file', None)
-  if (req.body.profilePicture)
-    console.log (`uploading image {req.body.profilePicture} start `)
+
+  if (req.body.profilePicture) 
+    console.log (`uploading image ${req.body.profilePicture} start `)
     try {
-      s3.upload_fileobj(photo_file, BUCKET, key)
-      
-
-      
-      photo.save()
+      const ReactS3Client = new S3(config)
+      ReactS3Client.uploadFile(file,req.body.profilePicture)     
+    }    
+    catch(error) {
+      return res.status(400).json(error); 
     }
-      
-    catch {
-      return res.status(400).json(err); 
-    }
-      
-
-
 }
 
 //keep, but doesn't do anything...
