+++ package.json
@@ -20,6 +20,7 @@
     "mongoose": "^6.7.3",
     "morgan": "^1.10.0",
     "react": "^18.2.0",
+    "react-aws-s3": "^1.5.0",
     "react-bootstrap": "^2.6.0",
     "react-dom": "^18.2.0",
     "react-input-emoji": "^4.3.0",
