--- src/components/UpdateUser/UpdatUser.css
@@ -1,28 +1,0 @@
-.updateContainer {
-    display:flex;
-}
-
-.updateContainer >label {
-    color:red
-}
-
-.infoInput {
-    border:1px solid #1165EB;
-    border-radius: 20px;
-    width:280px;
-    height: 46px;
-    color:#1165EB;
-    font-size: 14px;
-    text-align: center;
-}
-
-.submitBtn {
-    border:1px solid #1165EB;
-    border-radius: 20px;
-    background-color: #2F15D1;
-    width:280px;
-    height: 46px;
-    color:#ffffff;
-    font-size: 14px;
-    font-weight: bold;
-}
